Move providers to separate folder
